pyproject.toml
[project]
name = "hello-sf-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.12.3",
    "pydantic>=2.11.9",
    "requests>=2.32.5",
    "simple-salesforce>=1.12.9",
]
